requirements.txt
groq==0.20.0
python-dotenv==1.0.1
streamlit==1.43.2